==> 0. US activity data.Rmd <==
---
title: "US activity data"
author: "Dr. Chih-Yu Hung"
date: "2024-12-19"
output: html_document
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)
source("setup.R")

Fert_County_Crop <- read_csv("Outputs/3.0_Fertilizer_County_Crop_LCA.csv")
PRE_PET <- read_csv("Outputs/1.2_PET&PRE_County_30y.csv")
Topo <- read_csv("Outputs/2.2_Soil topo.csv")
NGS <- read_csv("Outputs/1.3_Non-growing season_30y.csv")
Texture <-read_csv("Outputs/2.1_Soil texture.csv")
```

## US activity data 

This file is used to prepare SiteData that will got to estimate N2O EF and SOC EF. 

## Direct N2O EF
### Direct N2O EF from synthetic fertilizer
```{r Direct EF }
#Prepare the P and PE for crops 
US_Direct <- Fert_County_Crop %>%
  left_join(PRE_PET, by = c("STATE", "COUNTY")) %>%
  #Some crops in some states are usually sown in fall so their growing season is between Oct-April or Nov-May
  mutate(
    P = case_when(
      CROP == "CANOLA"  & STATE %in% c("OKLAHOMA", "KANSAS", "CALIFORNIA", "TEXAS") ~ NGROW_PRE,  
      CROP %in% c("CAMELINA", "BARLEY") & STATE == "CALIFORNIA"               ~ NGROW_PRE,
      CROP == "RYE"     & STATE == "NORTH DAKOTA"                             ~ NGROW_PRE,
      str_detect(CLASS, regex("^winter", ignore_case = TRUE))       ~ NGROW_PRE,
      CROP == "TRITICALE" & DOUBLE == "Y"                           ~ NGROW_PRE,
      TRUE                                                          ~ GROW_PRE
    ),
    PE = case_when(
      CROP == "CANOLA"  & STATE %in% c("OKLAHOMA", "KANSAS", "CALIFORNIA", "TEXAS") ~ NGROW_PET,  
      CROP %in% c("CAMELINA", "BARLEY") & STATE == "CALIFORNIA"               ~ NGROW_PET,
      CROP == "RYE"     & STATE == "NORTH DAKOTA"                             ~ NGROW_PET,
      str_detect(CLASS, regex("^wint", ignore_case = TRUE))         ~ NGROW_PET,
      CROP == "TRITICALE" & DOUBLE == "Y"                           ~ NGROW_PET,
      TRUE                                                          ~ GROW_PET
    )
  ) %>%
  left_join(Topo, by = c("STATE" = "State")) %>%
  left_join(NGS, by = c("STATE" = "State", "COUNTY" = "County")) %>%
  left_join(Texture, by = c("STATE" = "State")) %>%
  select(STATE, COUNTY, YEAR, CROP, CLASS, DOUBLE, Fert_LCA_kg, P, PE, Topo, NGS, Frac_Fine, Frac_Medium, Frac_Coarse)


#write.csv(US_Direct, "Outputs/0.1_US_Direct.csv", row.names = F) 
```



## Indirect N2O
### volatilization
Volatilization require multiple soil properties and fertilizer data

```{r volatilization}
#Example for SiteData
#   RegionID:        suggest using HUP
#   ProvinceID:      Province or State ID
#   Year:            
#   CropID:          Name of crop
#   Crop_coef:       characteristic, Annual or Perennial
#   Fert_coef:       characteristic, Urea, Urea ammonia, Anhydrous ammonia, and Other
#   Method_ID_coef:  numeric, Broadcast (1) and Incorporated (2 or 3)
#   SoilpH:          numeric, -1 (<7.25) and 0 (>7.25)
#   SoilCEC:         numeric, or -1 (<250) and 0 (>250)
#   Climate          Temperate or Tropical
#   Percentage:      Percentage of soil condition (the combination of pH and CEC, sum of them must be 1)
#   Fraction:        Fraction of fertilizer
#   Proportion:      Proportion of application method
#   Fertilizer_Applied: The amount of applied fertilizer (kg)
#   Ecozone:         Ecozone for the Region, wet and dry based on annual PET and PRE!

#NOTE that these are all at STATE LEVEL
Soil_pH    <- read.csv("Outputs/2.3_Soil pH.csv")
Soil_CEC   <- read.csv("Outputs/2.4_Soil CEC.csv")
Fert_form  <- read.csv("Outputs/3.2_Fertilizer_form.csv")
Fert_method <- read.csv("Outputs/3.7 Fertilizer app method_STATE.csv")

#Prepare soil condition percentage 
Soil_percentage <- Soil_pH %>%
             left_join(Soil_CEC, by = "State") %>%
             mutate(Percentage = Fraction.x * Fraction.y) %>%
             select(State, SoilpH, SoilCEC, Percentage)

#Prepare fertilizer form, method at State level
Fert_form_method <- Fert_method %>%
                    mutate(Fert_coef = 
                           case_when(
                           Fertilizer_ID == 1 ~ "Urea",
                           Fertilizer_ID == 2 ~ "Urea ammonium nitrate",
                           Fertilizer_ID == 3 ~ "Anhydrous ammonia",
                           TRUE ~ "Other"
                           )) %>%
                    left_join(Fert_form, by = c("STATE", "CROP", "CLASS", "Fert_coef" = "FORM")) %>%
                    rename(Fraction = PERCENTAGE) %>%
                    select(-Fertilizer_ID,-REGION)

#Add annual PET and PRE to determine wet zone or dry zone
PRE_PET_y <- PRE_PET %>%
  mutate(PET_Y = GROW_PET + NGROW_PET,
         PRE_Y = GROW_PRE + NGROW_PRE) %>%
  select(STATE, COUNTY, PET_Y, PRE_Y)


#Prepare the SiteData for US volatilization
US_vola <- US_Direct %>%
  select(STATE:Fert_LCA_kg) %>%
  left_join(PRE_PET_y, by = c("STATE", "COUNTY")) %>%
  left_join(Soil_percentage, by = c("STATE" = "State")) %>%
  left_join(Fert_form_method, by = c("STATE", "CROP", "CLASS"))
   

#write.csv(US_vola, "Outputs/0.2_US_volatilization.csv", row.names = F) >1.2GB too large
#save(US_vola,file= "Outputs/0.2_US_volatilization.RData") #Only 10MB
```



### Leaching
 Leaching  only require precipitation data and fertilizer data 
 
```{r Leaching EF}
#Prepare the P and PE for crops 
US_leaching <- Fert_County_Crop %>%
  left_join(PRE_PET, by = c("STATE", "COUNTY")) %>%
   mutate(
    P = case_when(
      CROP == "CANOLA"  & STATE %in% c("OKLAHOMA", "KANSAS", "CALIFORNIA", "TEXAS") ~ NGROW_PRE,  
      CROP %in% c("CAMELINA", "BARLEY") & STATE == "CALIFORNIA"               ~ NGROW_PRE,
      CROP == "RYE"     & STATE == "NORTH DAKOTA"                             ~ NGROW_PRE,
      str_detect(CLASS, regex("^winter", ignore_case = TRUE))       ~ NGROW_PRE,
      CROP == "TRITICALE" & DOUBLE == "Y"                           ~ NGROW_PRE,
      TRUE                                                          ~ GROW_PRE
    ),
    PE = case_when(
      CROP == "CANOLA"  & STATE %in% c("OKLAHOMA", "KANSAS", "CALIFORNIA", "TEXAS") ~ NGROW_PET,  
      CROP %in% c("CAMELINA", "BARLEY") & STATE == "CALIFORNIA"               ~ NGROW_PET,
      CROP == "RYE"     & STATE == "NORTH DAKOTA"                             ~ NGROW_PET,
      str_detect(CLASS, regex("^wint", ignore_case = TRUE))         ~ NGROW_PET,
      CROP == "TRITICALE" & DOUBLE == "Y"                           ~ NGROW_PET,
      TRUE                                                          ~ GROW_PET
    )
  ) %>%
  select(STATE, COUNTY, YEAR, CROP, CLASS, DOUBLE, Fert_LCA_kg, P, PE)

#write.csv(US_leaching, "Outputs/0.3_US_leaching.csv", row.names = F) 
```
