==> 1.2 Growing season.Rmd <==
---
title: "1.2 Growing season"
author: "Dr. Chih-Yu Hung"
date: "2025-02-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)
source("setup.R")

Temp <- read_csv("Climate Normals/Monthly_Temp_perCounty_19812010.csv")

Pre <- read_csv("Climate Normals/Monthly_Pre_perCounty_19812010.csv")

Pet <- read_csv("Climate Normals/Monthly_Pet_perCounty_19812010.csv")


```


## Growing season for crop and States

This document will decide the growing season for crop and States. Please note that growing season in Canada is between May-Oct. However, US has double cropping. Growing season for double cropping could from April to September and Oct to March for winter crops. 


I'm using 10C (50F) Avg. month temperature as the first month of growing season. Two references below. 
https://www.epa.gov/climate-indicators/climate-change-indicators-growing-degree-days
https://extension.psu.edu/understanding-growing-degree-days


```{r determining growing season}
# If April’s mean temp >= 10, growing season = Apr–Sep; otherwise May–Oct
grow_season <- Temp %>%
  filter(month == 4) %>%
  mutate(
    start_month = if_else(mean_Temp >= 10, 4, 5),
    end_month   = start_month + 5            # Apr–Sep => 4–9; May–Oct => 5–10
  ) %>%
  select(County, State, start_month, end_month)


# Join Pre and Pet with growing-season info and sum over the growing season months
Pre_grow <- Pre %>%
  left_join(grow_season, by = c("County", "State")) %>%
  filter(month >= start_month & month <= end_month) %>%
  group_by(County, State) %>%
  summarise(GROW_PRE = sum(mean_Pre), .groups = "drop") %>%
  mutate(County = str_remove(County, " County$")) %>%
  mutate(State = toupper(State)) %>%
  rename("COUNTY" = "County","STATE" = "State")
  
Pet_grow <- Pet %>%
  left_join(grow_season, by = c("County", "State")) %>%
  filter(month >= start_month & month <= end_month) %>%
  group_by(County, State) %>%
  summarise(GROW_PET = sum(mean_Pet), .groups = "drop") %>%
  mutate(County = str_remove(County, " County$")) %>%
  mutate(State = toupper(State)) %>%
  rename("COUNTY" = "County","STATE" = "State")


# Join Pre and Pet with Non-growing-season info and sum over the NON growing season months
Pre_nongrow <- Pre %>%
  left_join(grow_season, by = c("County", "State")) %>%
  filter(month < start_month | month > end_month) %>%
  group_by(County, State) %>%
  summarise(NGROW_PRE = sum(mean_Pre), .groups = "drop") %>%
  mutate(County = str_remove(County, " County$")) %>%
  mutate(State = toupper(State)) %>%
  rename("COUNTY" = "County","STATE" = "State")
  
Pet_nongrow <- Pet %>%
  left_join(grow_season, by = c("County", "State")) %>%
  filter(month < start_month | month > end_month) %>%
  group_by(County, State) %>%
  summarise(NGROW_PET = sum(mean_Pet), .groups = "drop") %>%
  mutate(County = str_remove(County, " County$")) %>%
  mutate(State = toupper(State)) %>%
  rename("COUNTY" = "County","STATE" = "State")


#Combine the tables
PRE_PET <- reduce(list(Pre_grow, Pre_nongrow, Pet_grow, Pet_nongrow),
                  full_join, by = c("STATE", "COUNTY")) %>%
            mutate(across(where(is.numeric), ~round(.x,1)))

write.csv(PRE_PET,"Outputs/1.2_PET&PRE_County_30y.csv", row.names = FALSE)
```
