==> 2.4 Soil CEC.Rmd <==
---
title: "2.4 Soil CEC"
author: "Dr. Chih-Yu Hung"
date: "2025-04-25"
output: html_document
---

## Soil CEC data from SSURGO

The Soil Survey Geographic Database (SSURGO) collected by the National Cooperative Soil Survey over the course of a century. 
(https://www.nrcs.usda.gov/resources/data-and-reports/soil-survey-geographic-database-ssurgo)

They provide an R package, soilDB, to obtain data from SSURGO

```{r setup}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)
source("setup.R")

```


## Soil CEC

Every state has multiple mukey, which is the least survey unit in SSURGO. Each mukey has its unique CEC content. I'm going to use this to determine their soil CEC and based on their area to determine the percentage of soil fertility for volatilization. 


```{r Loop all state for soil texture}
# Initialize result table
# Initialize result table
cec_fraction_by_state <- data.frame()

# Loop through each state abbreviation
for (state_code in state_crosswalk$abbr) {
  
  # Get soil CEC (ph7, 0–30 cm)
  cec_data <- get_SDA_property(
    property = "cec7_r", 
    method = "Weighted Average", 
    WHERE = paste0("areasymbol LIKE '", state_code, "%'"), 
    top_depth = 0, 
    bottom_depth = 30
  )
  
  # Get area data for that state
  area_data <- get_mapunit_from_SDA(
    WHERE = paste0("areasymbol LIKE '", state_code, "%'")
  ) %>% select(mukey, muacres)
  
  # Join and classify
  cec_data_with_area <- left_join(cec_data, area_data, by = "mukey") %>%
    na.omit() %>%
    mutate(SoilCEC = ifelse(cec7_r < 25, -1, 0))
  
  # Summarize by pH class
  cec_area_fraction <- cec_data_with_area %>%
    group_by(SoilCEC) %>%
    summarise(area = sum(muacres, na.rm = TRUE), .groups = "drop") %>%
    mutate(abbr = state_code)  # keep abbreviation for now
  
  # Append to final result
  cec_fraction_by_state <- bind_rows(cec_fraction_by_state, cec_area_fraction)
}



# Add full state names from state_crosswalk$full
cec_fraction_by_state <- cec_fraction_by_state %>%
  left_join(state_crosswalk, by = "abbr") %>%
  group_by(full) %>%
  mutate(Fraction = area / sum(area)) %>%
  ungroup() %>%
  rename(State = full) %>% 
  mutate(State = toupper(State)) %>%
  select(State, SoilCEC, area,Fraction)

# Save to CSV if needed
#write.csv(cec_fraction_by_state, "Outputs/2.4_Soil CEC.csv", row.names = FALSE)


```
