==> 3.2 Fertilizer form.Rmd <==
---
title: "Fertilizer types"
author: "Dr. Chih-Yu Hung"
date: "2025-01-28"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(readxl)
library(magick)
```

## To obtain fertilizer types in different regions in the US. 

The methodologies to estimate fertilizer types (% for various types) are developed by Cao et al. 
Although it's not crop sepcific and only estimated the proportion to 2015 (and before). 

Only corn has the exception. 
Nitrogen fertilizer types of corn is estimated ny Cooter et al., 2012. 

Thanks to Prof. Lu provided their data. (Feb 4, 2025)


Ref. 
Cao, P., Lu, C., & Yu, Z. (2018).
Historical nitrogen fertilizer use in agricultural ecosystems of the contiguous United States during 1850–2015: application rate, timing, and fertilizer types. Earth System Science Data .https://doi.org/10.5194/essd-10-969-2018
Cooter, E.J, Bash, J.O. Benson, V. & Ran L. (2012).
Linking agricutural crop management and air quality models for regional to national-scale nitrogen assessments. Biogeosciens



```{r fertilizer form for corn}
Fertilizer_form_corn <- read_excel("Fertilizer/Fertilizer form corn.xlsx", 
    sheet = "Fertilizer form") 
#The percentages don't add up to 100% becasue it has manure. NEED to convert to Synthetic N fertilizer only

#Use Farm production region (FPR from 3.1 Fertilzation rate.Rmd)
FPR <- read.csv("Fertilizer/Farm Production Regions.csv") %>%
  mutate(STATE_NAME = toupper(State)) %>%
  select(STATE_NAME, Region) %>%
  rename(STATE = STATE_NAME,
         REGION = Region)
#Change WA to Mountain
FPR <- FPR %>%
  mutate(
    REGION = case_when(
      STATE == "WASHINGTON" ~ "Mountain",
      TRUE ~REGION
  ))

#calculate the percentage(with manure to N fertilizer only) and complete Corn table
Fertilizer_form_corn <- Fertilizer_form_corn %>%
  group_by(REGION) %>%
  mutate(PERCENTAGE = (PERCENTAGE/ sum(PERCENTAGE))) %>%
  ungroup() %>%
  mutate(CROP = "CORN") %>%
  left_join(FPR, "REGION") %>% 
  select(STATE, REGION, FORM, PERCENTAGE, CROP) %>%
  mutate(CLASS = "ALL CLASSES") # to match wheat

#write the data
#write.csv(Fertilizer_form_corn, "Outputs/Fertilizer_form_corn.csv",row.names = FALSE)
```

## Extract the figure in Cao and calculate the proportion. 

The figure is save to "Fertilizer"


```{r fertilizer form by Cao}
#
Fertilizer_Cao <- read_excel("Fertilizer/US_N_fertilizer_type_ratio_7_regions.xlsx")
# AnA = Anhydrous ammonia
# UR = urea
# AN = ammonium nitrate will be categorized to UAN


#read US's national Climate Assessment. Cao's study use this region
NCA_ref <- read.csv("Fertilizer/National Climate Assessment.csv", header = T) %>%
  mutate(State = toupper(State)) %>%
  rename(STATE = State)

#Wrangle the data
Fertilizer_form <- Fertilizer_Cao %>%
  filter(Year == 2015) %>%
  mutate(
    Urea = UR, #Cao used different symbols for fertilizers
    `Urea ammonium nitrate` = AN,
    `Anhydrous ammonia` = AnA,
    `Other N sources` = AqA + AS + NS + SN + CN + DAP + MAP + Aps
  ) %>%
  select(Region, Urea, `Urea ammonium nitrate`, `Anhydrous ammonia`, `Other N sources`) %>%
  rename(REGION = Region) %>%
  pivot_longer(
    cols = -REGION,
    names_to = "FORM",
    values_to = "PERCENTAGE"
  ) %>%
  left_join(NCA_ref, by = c("REGION" = "NCA_ID")) %>% #For the NCA and state names
  left_join(FPR, "STATE") %>% #FPR and states.
  select(STATE, REGION.y, FORM, PERCENTAGE) %>%
  rename (REGION = REGION.y)

#Imput fertilizer rate for names of other crops (excep corn)
Crop_noCORN <- read.csv("Outputs/3.1_Fertilization_Rate_State_Crop.csv",header=T) %>%
  filter(CROP != "CORN") %>%
  select(CROP, CLASS)

#Cross join to expand the data
Fertilizer_form_other <- Fertilizer_form %>%
  crossing(Crop_noCORN)

#save and write csv file
Fertilizer_form <- rbind(Fertilizer_form_corn, Fertilizer_form_other) %>%
  mutate(
    FORM = case_when(
      FORM == "Other N sources" ~ "Other",
      TRUE ~  FORM)
    )
#write.csv(Fertilizer_form, "Outputs/3.2_Fertilizer_form.csv", row.names = FALSE)
```
