==> 3.3 Crop area.Rmd <==
---
title: "3.3 Crop area"
author: "Dr. Chih-Yu Hung"
date: "2025-01-30"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
```

## Crop area

The crop area data is from Cropland Data Layer (https://croplandcros.scinet.usda.gov/). 
Thanks to Daniel Aydalla (PIRD, AFOLU) extract the crop area by applying county data. The raster data provided 30 x 30 m pixel and we converted it to ha. 

```{r crop area}
Crop_area <- read.csv("Crop/USAcrops_2018_2022_perCounty_Merge.csv")

# Initial wrangling.  
Crop_area <- Crop_area %>%
     separate(County_State, into = c("County", "State"), sep = " - ") %>% #Seperate state and county to two columns.
     #mutate(Year_Group = "2018_2022") %>%
     select(-Background) %>% #Background is the sea or the Great Lakes
     mutate(State = toupper(State))
     #group_by(County,State,Year_Group) %>%
     #summarize(across(where(is.numeric), mean, na.rm = TRUE)) %>%  # Compute mean for all numeric columns
     #ungroup()  # Remove grouping
  


#non agriculture land
# Non_agr <- c("Pasture_Grass",	"Forest",	"Shrubland",	"Barren", "Christmas_Trees",
#              "Other_Tree_Crops", "Clouds_No_Data",	"Developed",	"Water",	"Wetlands",
#              "Nonag_Undefined",	"Open_Water",	"Perennial_Ice_Snow",
#              "Developed_Open_Space",	"Developed_Low_Intensity",	"Developed_Med_Intensity",
#              "Developed_High_Intensity",	"Barren_1",	"Deciduous_Forest",	"Evergreen_Forest",
#              "Mixed_Forest",	"Shrubland_1",	"Grassland_Pasture",	"Woody_Wetlands",
#              "Herbaceous_Wetlands")

Other_agr <- c("Tobacco","Other_Crops", "Sod_Grass_Seed", "Alfalfa","Other_Hay_Non_Alfalfa","Switchgrass",
               "Fallow_Idle_Cropland",	"Aquaculture")

Vegetables <- c( 
  "Chick_Peas", "Lentils", "Dry_Beans", "Clover_Wildflowers",
  "Sweet_Corn", "Onions", "Cucumbers", "Tomatoes", "Carrots", "Asparagus", "Garlic", "Vetch",
  "Broccoli", "Peppers", "Lettuce", "Pumpkins", "Cauliflower", "Celery", "Radishes", "Cotton",
  "Turnips", "Eggplants", "Gourds", "Cabbage", "Squash", "Pop_or_Orn_Corn","Olives", "Greens",
  "Herbs", "Mint", "Hops","Sugarbeets"
)

Fruits <- c(
  "Misc_Vegs___Fruits", "Oranges","Pears",
  "Watermelons", "Caneberries", "Cherries", "Peaches", "Apples", "Grapes", "Citrus",
  "Pomegranates", "Nectarines", "Plums", "Strawberries", "Apricots", "Blueberries",
  "Cranberries", "Honeydew_Melons", "Cantaloupes", "Prunes", "Avocados","Sugarcane"
)

Grains_exclude <- c(
  "Rice", "Other_Small_Grains", "Oats", "Millet", "Speltz", "Buckwheat", "Potatoes", "Sweet_Potatoes"
)

Oilseeds_exclude <- c(
  "Sunflower", "Flaxseed", "Safflower", "Rape_Seed", "Mustard","Pistachios","Peanuts","Pecans","Almonds",
  "Walnuts"
)

Target_crop <- c("Barley", "Camelina", "Canola", "Corn", "Peas", "Rye", "Sorghum", "Soybeans", "Triticale",
               "Durum_Wheat", "Spring_Wheat", "Winter_Wheat"
)

# Extract column names that start with "Dbl_"
Dbl_crop_columns <- grep("^Dbl_", colnames(Crop_area), value = TRUE)
# Exclude columns containing "Cantaloupe" or "Lettuce_Cotton"
Dbl_crop <- Dbl_crop_columns[!grepl("Cantaloupe|Lettuce_Cotton", Dbl_crop_columns)]
Dbl_crop_exclude <- Dbl_crop_columns[grepl("Cantaloupe|Lettuce_Cotton", Dbl_crop_columns)]

Agr_area <- Crop_area %>%
               mutate(
               Fruits = rowSums(select(., all_of(Fruits)), na.rm = TRUE),
               Vegetables = rowSums(select(., all_of(Vegetables)), na.rm = TRUE),
               Grains_exclude = rowSums(select(., all_of(Grains_exclude)), na.rm = TRUE),
               Oilseeds_exclude = rowSums(select(., all_of(Oilseeds_exclude)), na.rm = TRUE),
               Other_agr = rowSums(select(., all_of(Other_agr)), na.rm = TRUE)) %>%
               select(
               State, County, Year, all_of(Target_crop), Fruits, Vegetables, 
               Grains_exclude,Oilseeds_exclude, Other_agr,
               all_of(Dbl_crop),all_of(Dbl_crop_exclude)) %>%
               mutate(
               Total_Dbl_area = rowSums(select(.,all_of(Dbl_crop)), na.rm = TRUE),    
               Total_Ag_area =  rowSums(select(.,where(is.numeric)), na.rm = TRUE),
               Dbl_percentage = round(Total_Dbl_area/Total_Ag_area *100,1))  

#Below may be outdated Feb 11,2025
#Consider the area of Dbl crops, I will deal with major Dbl crop only (they have 92.3% of Dbl crops)
#This include "Winter wht+soybean"(75%) , "Winter wht +corn"(7.2%)  "Winter wht + Sorghum"(6.8%),  "Triticale +corn"(3.1%) Jan. 31. 
          
#write.csv(Agr_area,"Outputs/3.3_Agriculture_area.csv",row.names = FALSE)
```

## Deal with the double crop area. 

The double crop area should be duplicated. 

```{r Double crop area}

# Extract column names that start with "Dbl_Crop_"
Dbl_crop_columns <- grep("^Dbl_Crop_", colnames(Agr_area), value = TRUE)


# Extract unique crops, ensuring full wheat names are preserved
unique_crops <- Dbl_crop_columns %>%
  gsub("^Dbl_Crop_", "", .) %>%  # Remove "Dbl_Crop_" prefix
  strsplit("_") %>%  # Split into crop pairs
  unlist() %>%  # Flatten the list
  unique()  # Keep unique crop names


# Initialize new crop columns with zeros
for (crop in unique_crops) {
  Agr_area[[paste0("Dbl_", crop)]] <- 0  # Initialize with 0 for summation
}


# Loop through each Dbl_Crop_ column and correctly sum values across crops
for (dbl_col in Dbl_crop_columns) {
  # Extract the two crops from the column name
  crops <- gsub("^Dbl_Crop_", "", dbl_col) %>% 
    strsplit("_") %>%
    unlist()
  
  # Sum the values into their respective new columns
  for (crop in crops) {
    Agr_area[[paste0("Dbl_", crop)]] <- Agr_area[[paste0("Dbl_", crop)]] + Agr_area[[dbl_col]]
  }
}


# Remove the orginal columns for double crops
Agr_area <- Agr_area %>%
 rename(Dbl_Winter_Wheat = Dbl_WinWht, Dbl_Durum_Wheat = Dbl_Durum) %>%
 select(-starts_with("Dbl_Crop_"), -Dbl_Wht, -Total_Dbl_area, -Dbl_percentage, - Total_Ag_area)
 

#Pivot to long table
Agr_area_long <- Agr_area %>%
  pivot_longer(
    cols = -c(State, County, Year),
    names_to = "CROP",
    values_to = "AREA"
  ) %>%
  mutate(CROP = toupper(CROP))

```

### I add a variable to classify if the area is for double crop and add CLASS to match CENSUS format.

NOTE: Crop area in the DISTRICT of COLUMBIA is removed. 

```{r Dbl crop area variable and CENSUS format}

# Pivot the data so that crop names become a single column and values remain under "Value"
Agr_area_final <- Agr_area_long %>%
  # Add the "DOUBLE" column: "Y" if Crop contains "DBL_", otherwise "N"
  mutate(DOUBLE = ifelse(grepl("^DBL_", CROP), "Y", "N")) %>%
  # Remove "Dbl_" prefix from Crop names
  mutate(CROP = gsub("^DBL_", "", CROP)) %>%
  # Add CLASS column based on rules
  mutate(CLASS = case_when(
    CROP == "PEAS" ~ "GREEN",
    CROP == "DURUM_WHEAT" ~ "SPRING, DURUM",
    CROP == "SPRING_WHEAT" ~ "SPRING, (EXCL DURUM)",
    CROP == "WINTER_WHEAT" ~ "WINTER",
    TRUE ~ "ALL CLASSES"  # Default to NA for other crops
  )) %>%
  mutate(CROP = gsub("DURUM_|SPRING_|WINTER_","",CROP)) %>% #remove the prefixes
  rename(STATE= State, COUNTY = County, YEAR = Year) %>%
  mutate(COUNTY = COUNTY %>%
           str_remove(regex("\\s+County$",ignore_case = TRUE)) %>%
           str_trim())  %>%
  filter(STATE != "DISTRICT OF COLUMBIA")



  
#write.csv(Agr_area_final, "Outputs/3.3_Area_County_Crop.csv", row.names = FALSE)

```


### Aggregated data for State only

This is for Antoine's work

```{r State aggregated area data}
# Agr_area_state <- Agr_area_final %>%
#   group_by(STATE, YEAR, CROP, CLASS, DOUBLE) %>%
#   summarise(State_area = sum (AREA, na.rm = TRUE)) %>%
#   group_by(STATE, CROP, CLASS, DOUBLE) %>%
#   summarise(Avg_State_area = mean(State_area, na.rm = TRUE)) %>%
#   ungroup()
# 
#   
# write.csv(Agr_area_state, "Outputs/3.3_Area_State_Crop_Antoine.csv", row.names = FALSE)

```
